backend/requirements.txt
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask==3.0.1
SQLAlchemy==2.0.25
configparser==5.0.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pyyaml==6.0
requests==2.27.1
#uWSGI==2.0.20
wheel==0.37.1
flask-marshmallow==0.15.0
alembic==1.11.1
apispec-webframeworks==0.5.2
apispec==6.3.0
Werkzeug==3.0.1
flask-swagger-ui==4.11.1
marshmallow==3.19.0
pytz==2023.3
Jinja2==3.1.2
